refactored session logic
