+++ src/app/layout.tsx
@@ -4,41 +4,13 @@ import '@mantine/core/styles.css';
 import '@mantine/notifications/styles.css';
 import './globals.css';
 import { ColorSchemeScript, MantineProvider } from '@mantine/core';
-import { User } from './sessions/AuthContext';
-import { AuthProvider } from './sessions/AuthProvider';
-import { filterStandardClaims } from 'next-firebase-auth-edge/lib/auth/claims';
-import { Tokens, getTokens } from 'next-firebase-auth-edge';
-import { cookies } from 'next/headers';
-import { serverConfig } from '../auth-config';
-
-const getUserSession = ({ decodedToken }: Tokens): User => {
-  const {
-    uid,
-    email,
-    picture: photoURL,
-    email_verified: emailVerified,
-    phone_number: phoneNumber,
-    name: displayName,
-    source_sign_in_provider: signInProvider, // FIXME: this exposes the password in the emulator
-  } = decodedToken;
-
-  // This abstracts all of the custom claims into a single object
-  const customClaims = filterStandardClaims(decodedToken);
-
-  return {
-    uid,
-    email: email ?? null,
-    displayName: displayName ?? null,
-    photoURL: photoURL ?? null,
-    phoneNumber: phoneNumber ?? null,
-    emailVerified: emailVerified ?? false,
-    providerId: signInProvider,
-    customClaims,
-  };
-};
-import ProtectedLayout from '@/components/ProtectedLayout';
 import { Notifications } from '@mantine/notifications';
-import '@mantine/notifications/styles.css';
+import { AuthProvider } from '@/sessions/AuthProvider';
+import { getUserSession } from '@/sessions/getUserSession';
+import { getTokens } from 'next-firebase-auth-edge';
+import { cookies } from 'next/headers';
+import { serverConfig } from '@/auth-config';
+import ProtectedLayout from '@/components/ProtectedLayout';
 
 const geistSans = localFont({
   src: './fonts/GeistVF.woff',
@@ -61,12 +33,12 @@ export default async function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  // This will be passed into the AuthProvider to allow session management
-  // TODO: Claims should be put into the session as well to manage page access with middleware, the tokens themselves will be used to authenticate with the server
   const tokens = await getTokens(await cookies(), {
     apiKey: process.env.NEXT_PUBLIC_API_KEY!,
     ...serverConfig,
   });
+
+  // Create user session based on the tokens and pass into the Context API
   const user = tokens ? getUserSession(tokens) : null;
 
   return (
